Financial_agent_with_phidata/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi[standard]
uvicorn
groq
openai
google-generativeai
sqlalchemy